# api/swagger.yaml
---
swagger: "2.0"
info:
  description: "\nТобі потрібно реалізувати сервіс з АРІ, який дозволить: \n- дізн\
    атись поточний курс біткоіну (BTC) у гривні (UAH)\n- підписати емейл на отрима\
    ння інформації по зміні курсу\n- запит, який відправить всім підписаним корист\
    увачам актуальний курс. \n\nДодаткові вимоги:\n1. Сервіс має відповідати описа\
    ному ниже АРІ. <i>NB Закривати рішення аутентифікацією не потрібно</i>. \n2. В\
    сі данні, для роботи додатку повинні зберігатися в файлах (підключати базу дан\
    них не потрібно). Тобто, потрібно реалізувати збереження та роботу з даними (н\
    априклад, електронними адресами) через файлову систему.\n3. В репозиторії пови\
    нен бути Dockerfile, який дозволяє запустити систему в Docker. З матеріалом по\
    \ Docker вам необхідно ознайомитись самостійно.\n4. Також ти можеш додавати ко\
    ментарі чи опис логіки виконання роботи в README.md документі. Правильна логік\
    а може стати перевагою при оцінюванні, якщо ти не повністю виконаєш завдання.\n\
    \nОчікувані мови виконання завдання: PHP, Go, JavaScript (Node.js).  Виконуват\
    и завдання іншими мовами можна, проте, це не буде перевагою.\nВиконане завданн\
    я необхідно завантажити на GitHub (публічний репозиторій) та сабмітнути викона\
    не завдання в гугл-форму.\n\nТи можеш користуватися усією доступною інформаціє\
    ю, але виконуй завдання самостійно. \nУспіхів! "
  version: "1.0.0"
  title: "GSES2 BTC application"
host: "gses2.app"
basePath: "/api"
tags:
  - name: "rate"
    description: "Отримання поточного курсу BTC до UAH"
  - name: "subscription"
    description: "Робота з підпискою"
schemes:
  - "http"
paths:
  /rate:
    get:
      tags:
        - "rate"
      summary: "Отримати поточний курс BTC до UAH"
      description: "Запит має повертати поточний курс BTC до UAH використовуючи бу\
        дь-який third party сервіс з публічним АРІ"
      operationId: "rate"
      produces:
        - "application/json"
      parameters: []
      responses:
        "200":
          description: "Повертається актуальний курс BTC до UAH"
          schema:
            type: "integer"
        "400":
          description: "Invalid status value"
  /subscribe:
    post:
      tags:
        - "subscription"
      summary: "Підписати емейл на отримання поточного курсу"
      description: "Запит має перевірити, чи немає данної електронної адреси в пот\
        очній базі даних (файловій) і, в разі її відсутності, записувати її. Пізні\
        ше, за допомогою іншого запиту ми будемо відправляти лист на ті електронні\
        \ адреси, які будуть в цій базі. "
      operationId: "subscribe"
      consumes:
        - "application/x-www-form-urlencoded"
      produces:
        - "application/json"
      parameters:
        - name: "email"
          in: "formData"
          description: "Електронна адреса, яку потрібно підписати"
          required: true
          type: "string"
          x-exportParamName: "Email"
      responses:
        "200":
          description: "E-mail додано"
        "409":
          description: "Повертати, якщо e-mail вже є в базі даних (файловій)"
  /sendEmails:
    post:
      tags:
        - "subscription"
      summary: "Відправити e-mail з поточним курсом на всі підписані електронні по\
        шти."
      description: "Запит має отримувати актуальний курс BTC до UAH за допомогою third-party\
        \ сервісу та відправляти його на всі електронні адреси, які були підписані\
        \ раніше.  "
      operationId: "sendEmails"
      produces:
        - "application/json"
      parameters: []
      responses:
        "200":
          description: "E-mailʼи відправлено"